requirements.txt
# Core ML and Data Processing
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.7.2
seaborn>=0.12.2
shap>=0.42.1
xgboost>=1.7.6
lightgbm>=4.0.0
catboost>=1.2.2; platform_system!="Windows"
catboost>=1.2.2; platform_system=="Windows" --only-binary :all:
scipy>=1.11.3
ydata-profiling>=4.5.1
pandera>=0.17.1

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.1
torch>=2.0.0
torchvision>=0.15.2
transformers>=4.30.0

# MLflow and Experiment Tracking
mlflow>=2.8.0
optuna>=3.3.0
hyperopt>=0.2.7

# Airflow and Workflow Management
apache-airflow>=2.7.1
apache-airflow-providers-apache-spark>=4.1.3
apache-airflow-providers-databricks>=4.1.1
apache-airflow-providers-amazon>=8.15.0
apache-airflow-providers-http>=4.5.0
apache-airflow-providers-postgres>=5.6.0
apache-airflow-providers-redis>=3.3.1
apache-airflow-providers-slack>=7.3.0

# WebSocket and Real-time Communication
websockets>=11.0.3
aiohttp>=3.8.5
fastapi>=0.100.0
uvicorn>=0.23.0

# Database and Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
redis>=4.6.0
pymongo>=4.5.0
boto3>=1.28.0
tenacity>=8.2.3

# Monitoring and Logging
prometheus-client>=0.17.0
grafana-api-client>=2.1.0
python-json-logger>=2.0.7
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1
mocha>=0.1.0
chai>=0.1.0
ws>=8.13.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
tqdm>=4.65.0
joblib>=1.3.2 